requirements.txt
chromadb==0.5.0
streamlit==1.38.0
beautifulsoup4==4.12.3
pdfplumber==0.11.4
plotly==5.24.0
pandas==2.2.2

langchain-core==0.2.37
langchain-openai==0.1.23
langchain-community==0.2.5
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langgraph==0.2.16
